Return integer dtype index vectors from BQM.to_numpy_vectors

+++ dimod/binary/binary_quadratic_model.py
@@ -1827,8 +1827,8 @@ class BinaryQuadraticModel:
             icol.append(label_to_idx[v])
             qdata.append(bias)
 
-        irow = np.asarray(irow)
-        icol = np.asarray(icol)
+        irow = np.asarray(irow, dtype=np.int64)
+        icol = np.asarray(icol, dtype=np.int64)
         qdata = np.asarray(qdata)
 
         if sort_indices:
